requirements.txt
alembic==1.5.0
click==7.1.2
ecdsa==0.18.0
fastapi==0.63.0
greenlet==2.0.2
h11==0.14.0
Mako==1.2.4
MarkupSafe==2.1.2
pyasn1==0.4.8
pydantic==1.10.4
PyJWT==2.6.0
python-dateutil==2.8.2
python-editor==1.0.4
python-jose==3.1.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.0
starlette==0.13.6
typing_extensions==4.4.0
uvicorn==0.13.2
